Implement second controller to watch configMap and reconcile.
Data for configs is read directly from configMap without need of mounting and reading files (which add latency)

Old code is not removed but commented.
The new controller watches ConfigMap from extension namespace - so we need to provide it with downward API but maps to all extensions to all shots.